requirements.txt.txt
streamlit
langchain-core
langchain_community
langchain
langchain-qdrant
qdrant-client
langchain_huggingface
python-dotenv
langchain-openai
